+++ src/lib/settings.ts
@@ -1,7 +1,7 @@
 import { writable } from 'svelte-persistent-store/dist/local'
 
 export interface GameSettings {
-  kanaType: 'hiragana' | 'katakana' | 'both'
+  kanaType: 'hiragana' | 'katakana' | 'both',
 }
 
 export const settings = writable<GameSettings>('game-settings', {
